AT-3268 fix `SearchForUsagesRunnable` to mark the search as in progress when creating the usage view
At some point in the past, it was changed for the usage view to have search in progress set to `false` by default on creation. This was done to enable refactoring targets that have no usages, but it meant that the "is search in progress" flag was always false for when searches were actually in progress.

It seems logical to manually set the search to in progress inside `SearchForUsagesRunnable`, given that:
* The API for creating a usage view is centered around initial `Usage`s that will pre-populate the view
* It's not required for callers to use a background task to append usages

GitOrigin-RevId: 26b685ee5d1df167bd0ba4973346b26387ef7130