[terminal] IJPL-208919 Migrate the blocks model to absolute offsets
This is harder than it would seem.

The model relies on document listeners because it needs
to track the actual changed range, which the output model
currently doesn't support (it doesn't compare
the before/after texts). But using absolute indices
means we need the model to be in a consistent state,
specifically, we need the trimmed character count to reflect
the actual state.

So we need to actually update one property before
even changing the document in the output model.
We can't update all of them, because, for instance,
the new line count isn't known until we update.

Even worse, we need to deal with the highlighting
model first, because it relies on the "old" indexes.

Maybe we can simplify it when we have smarter listeners
in the model.

GitOrigin-RevId: 3f19df1554afc4c4889dec1df7a57d29b732a8d8