[Gradle][VersionCatalog] IDEA-370426 generate version catalog accessors relying on catalog locations from Gradle sync
IDEA generates SyntheticVersionCatalogAccessor classes to provide navigation from a version catalog usage to a particular element of the catalog. Previously, GradleDslVersionCatalogHandler had been generating the accessors relying only on the ProjectBuildModel by Android Gradle DSL that parses Gradle settings files and TOML files. The lack of this approach is in unability of the parser to cover some tricky cases.

Now, IDEA gets more reliable information about version catalog locations from the Gradle at sync and stores it in WSM. This commit enables using this data when generating the accessors. Thus, accessors could be generated not only for the catalogs the Android parser knows, but also for the catalog locations received from Gradle.

Note: in the case of dynamically added subproject, the full completion of a version catalog reference doesn't work. Only step by step. Created a test and an issue for that - IDEA-380354.

Tests:
* org.jetbrains.plugins.gradle.dsl.versionCatalogs.GradleVersionCatalogsCompletionTest.testCompletionLongInDynamicallyAddedSubproject
* org.jetbrains.plugins.gradle.dsl.versionCatalogs.GradleVersionCatalogsFindUsagesTest.testDynamicallyAddedSubprojectAndCustomToml
* org.jetbrains.plugins.gradle.dsl.versionCatalogs.GradleVersionCatalogsRenameTest.test renaming a library in a custom TOML, for dynamically added subproject
* org.jetbrains.plugins.gradle.dsl.versionCatalogs.GradleVersionCatalogsResolveTest.testNavigationFromDynamicallyAddedSubprojectToCustomToml

GitOrigin-RevId: 09526dd2f98970552811bf07bd09e626e09eb5c0