[UI Inspector] IJPL-211726 Redesign DimensionsComponent
The old design was confusing: the insets
and border are the same in most cases.
Yet, it was designed, as if they add up.
Moreover, the rectangles created the false
impression that the size doesn't include
the insets, but it does.

The new look may be not so impressive,
but it reflects the actual state better.

GitOrigin-RevId: c78e384a94ea114c16778b5bf95bb0144d3ee8b8