[platform UI] IJPL-208221 Fix tool window background image
The issue is caused by two things in XNextIslandHolder:
unwrapping the Graphics and setting NO_BACKGROUND.
Both are needed only for island themes, so let's move
them to InternalUICustomization. For the NO_BACKGROUND
property there's a function already, but for unwrapping/wrapping
there doesn't seem to be anything exactly like that.
So let's create a new function for this case.

GitOrigin-RevId: 142b1fb416aeabf0d2c2abd2982c758f6ad25106