[terminal] IJPL-208919 Refactor terminal output model changing
Reimplement updateContent using doReplaceContent,
so we only have major updating / trimming in one place.

Extract trimming to changeDocumentContent.

Make the hyperlink tests more robust by using EOL
as much as possible, to avoid the question of
whether a filter should be reapplied to a line
if the only change to that line is the addition of the EOL.

(cherry picked from commit 867bc4e80eb808a735dcec83da5825abfd635bc1)

IJ-CR-177708

GitOrigin-RevId: 891ee452e513d2e1c4ef0b3700b9a25de05ef659