[terminal] IJPL-208919 Split terminal output model events
To make implementation easier, report every change to the document
as a separate event (and trim after every change).

This will make "service" updates (like adding empty lines or spaces)
separate events, but it's more of a good thing, because it'll make
the real events separate and because it makes the code simpler.

GitOrigin-RevId: 4f914284c4098cefe2b03494eee10dcce6500c1d