[terminal] IJPL-208919 Implement TerminalContentChanged
Now we have separate events for trimming and regular updates.
We also have more information about every update,
and it's more consistent: previously, it would "trim" the start
offset to "coerce" it into the document range.
Now the actual change is reported while the new text is still there.

Add tests that check that every change is reported
and that the state is consistent on every event.

GitOrigin-RevId: 80f36ecaf7f0326f6ba24529146c4ff23778a408