[terminal] IJPL-212661 Fix terminal tabs restoring
It looks like there was a race and our "close project" heuristic didn't work because we run the check in the background thread.
Now there should be even more guarantees that it won't run on project close: we run the RPC call in the scope of the `TerminalToolWindowTabsManager`. It should be closed at this moment.

GitOrigin-RevId: f30bc9e2f4676080414bcea02688783fefbf74ff