[JEWEL-1028] Add toolwindow bg to GlobalColors
This change is important for IntelliJ UI as we did not have a
simple and portable way to access the toolwindow colour until
now.

This being Jewel, we of course have it in standalone too, not
only in bridge, but the colour is meaningless in standalone,
so it's unspecified. The older GlobalColors factory functions
have been deprecated and hidden — I checked this does not
cause source breakages.

closes https://github.com/JetBrains/intellij-community/pull/3275

GitOrigin-RevId: d26589699f46d9ed19a1db2776228f59c46e226c