[JEWEL-1052] Cleanup deprecated APIs
This marks a bunch of deprecated APIs as hidden, solving several
issues with deprecation notices appearing to users. There are no
source level breakages (at least in our usages), and the binary
compat is still guaranteed as the symbols are still in the class
files.

The other notable change is that we've un-deprecated Dropdown
and marked it as experimental, since it still has uses that
have no proper replacement.

* DataProviderNode is now annotated as internal (and no longer
  deprecated) as it was always supposed to be
* MenuItemColors styling APIs using the deprecated-to-hide
  MenuItemState APIs have been deprecated-to-hide, too
* Many deprecated API with source compatible overloads have been
  hidden (note: source compat can always be broken in source, but
  I've only done it on a few cases where it should not be an issue
  for anyone, and on APIs deprecated since at least 0.29)
* ColorExtensions' deprecated APIs have not been hidden since they
  have a different behaviour compared to the newer ones