Don't notify the IDE when using the Compose Markdown
... renderer in Android Studio.

Since, JCef is not supported in Android Studio there is no use in
notifying our users that the compose panel provider is used or polluting
our logs.

This small diff is preferred rather than setting the preferredProvider
initially to the compose provider as that introduces a cyclical
dependency between the intellij.markdown intellij.markdown.compose.preview
modules.

Bug: 451595711
Change-Id: I1f57d8fd2319a759c037fde3b5ce9a31eb282883