[rdct] IJPL-199766: Implement file-based settings sync
Some settings are only used in the monolith and on the backend. In this case, PersistentStateComponents aren't loaded on the frontend so the RdSettingsController can't be used to sync these settings. They can only be synced between the frontend and backend as raw files.

(cherry picked from commit 226200051312e64123941aeacee626883b6c8b03)

GitOrigin-RevId: 36bf9d4daed5d1e7358ff9d2a6de1fd664726b3e