[python] Don't depend on `types.UnionType` as inferred type for `Type1 | Type2` in PyTypeAssertionEvaluator
In the recent update of Typeshed they have updated `types.UnionType.__or__` to return
`UnionType | Any` instead of the standalone `UnionType`, breaking type narrowing.
Overall, this type seems more like an implementation detail of typing in CPython,
it's better to rely on an explicit notation accepted as the second argument of
`isinstance`/`issubclass`.

See the tests affected by the Typeshed changes:
* PyTypingTest#testBitwiseOrUnionIsSubclass
* PyTypingTest#testBitwiseOrUnionIsInstance
* PyTypingTest#testWalrusIsSubclass

GitOrigin-RevId: 538f9f7f73b6a646b449f2d6ca218b308f08ebca