[JEWEL-1051] Fix Metalava CI checks, add gen'd theme sanity checks
This fixes the issues we've been having with PRs that introduce changes
in a release, and also gives us a way to deal with the platform removing
deprecated and unused APIs in their usual cleanup cycles.

The solution is using Metalava baselines for any changes mid-cycle; we
write all changes there and, when it's time to release, we validate and
store the changes for that release. While this is not a perfect solution
it at least allows us to see those baseline changes in the PRs. Then we
can at least review them.

The release and PR documentations were updated accordingly.

This also updates the generated themes and adds a sanity check test that
should ensure they get updated when the major platform version is also
bumped (and, that we don't mistakenly cherry-pick their changes to an
older release branch).

Note: this also updates ktfmt's IJ plugin settings since the new plugin
version changed the storage format slightly.

closes https://github.com/JetBrains/intellij-community/pull/3276

GitOrigin-RevId: 5127f3429d3a16710797c6e7e279abd0dcfbd21e