[platform UI] IJPL-156745 Introduce the wayland.true.popups key
There are pros and cons to using true popups,
so we need a key to experiment with them.

For now location restoring is disabled (even though it's
one of the main benefits of true popups)
because we first need to ensure location correctness.
Otherwise, an attempt to restore the location
may cause the popup to immediately disappear,
making it permanently unusable.

GitOrigin-RevId: 7597f770f339fc8f7dfbb14d4cf467f0d8e09035