Swallow java.lang.Error in BatchSpanProcessor
We've stumbled onto an (overall unrelated) `java.lang.ClassNotFoundException` @ BatchSpanProcessor.kt:99 while compiling the IJ monorepo as a Kotlin user project. This killed the infinite loop responsible for processing BatchSpanProcessor's channels and, AFAIU, left a user of BatchSpanProcessor hanging. The overall result was a 3-hours-long timed-out TC run.

I'm not sure if this change actually helps, but hopefully it'll work analogously to try-catch-finally blocks in BatchSpanProcessor that _do_ already catch Throwable (and, by extension, that it'll work properly).

GitOrigin-RevId: 94bafec84e86400e85e1e7a2c253e5ba0f3b4a44