[platform UI] IJPL-156745 Make true popups resizable on Wayland
Now that we have extracted the implementation from the listener,
it isn't that hard. We just need to use the new pointer API
to track the movement. Also, there's a special handling
for the mouse released event because it arrives on Wayland
no matter what and closes the popup immediately after resizing,
so we have to consume it to avoid this.

GitOrigin-RevId: 6e93b976d570d106e8482ee7b6fa258597e820cd