[terminal] IJPL-211122 Refactor: introduce TerminalCommandBlock interface
It is the replacement for existing `TerminalOutputBlock`.
Design it with a possibility to have different block types. Currently, there is only a single `TerminalCommandBlock`, but we can possibly add other types (for example, different decorations with content).
To make it possible, extract `TerminalBlockBase` with properties any block should have. And use it in most places.

GitOrigin-RevId: 3793387365725d65c484fd07c5c383668a5331ba