[terminal] IJPL-211122 Refactor: do not use sealed interfaces in blocks model
It is good to use `sealed` when you are sure that there will be no new values.
But in this case it is not true. So, if the new implementation added, it may break the binary compatibility for clients that use `when` statement.

(cherry picked from commit b11d62975e182731f7fb09f3f45b46a031f8d7c6)

IJ-CR-179105

GitOrigin-RevId: b04b1afa41b2131a03342419ebd4006aec4fa4f1