[terminal] IJPL-213148 Fix the highlighting model update on clear
When the output model is cleared, the offsets are changed
in an unusual way, effectively invalidating absolute offsets
that are otherwise pretty reliable.

This caused the highlighting model to not clear some
outdated highlightings because it was confused by indices.

Fix by creating a special clear() function that resets
everything in a consistent manner. And update the tests.

GitOrigin-RevId: f346eb34eac5f5f71dd45406b405211da4b08d68