[JEWEL-1061] Fix split button chevron getting squashed
This fixes an issue with the secondary content (the chevron)
getting squashed in split buttons, when there isn't enough
space to show both the primary and secondary contents fully.

As a bonus, this also fixes:
 * The divider height is no longer capped, but scales
   correctly with the button's height
 * Colours were all sorts of wrong in standalone mode for
   buttons, especially in the disabled state (they are fine
   in bridge mode)
 * The semantics role of the secondary action is now the same
   as in Swing (button)
 * The modifier passed to split buttons is now applied in
   the right place, and the internal layout is simplified

The buttons showcase is also improved, now showing a space-
constrained split button that did not use to work correctly,
and putting all the content in flow rows so it scales better
as the window is resized. The visual appearance is also
tweaked to better represent the information hierarchy.