[python] (PY-84870) disable buttons on the target custom tool wizard in case of tool validation errors
Originally the "Create" button on the Target Custom Tool Wizard (it is used only for Python now) was always active not depending on the tool internal state. registerStateChangedCallback was added to the CustomToolLanguageConfigurable interface to notify controls of changes inside the tool form.

CustomToolLanguageConfigurable was marked as Internal to align with TargetCustomToolPanel.


Merge-request: IJ-MR-179005
Merged-by: Vitaly Legchilkin <Vitaly.Legchilkin@jetbrains.com>

GitOrigin-RevId: dfcf3fe18f3cd283577c0c5afda2be0423105f1f