[gradle] IJPL-213214 support possible remote (eel-based) `javaHome`.
Gradle uses `io.File` as `javaHome` so it doesn't have filesystem info (it is always local).

We set `File("/path/WSL/java")` for WSL (or Docker) on Windows and `File` replaces `/` with `\` (`File.separator`@Windows) so it becomes `File("\path\WSL\java")` following `path` call canonicalizes it to `c:\path\WSL\java` and we end up this nonsense passed as `homePath`.

For the Unix we might have valid path, but local instead of remote: i.e: "/foo" instead of "\\docker\foo"

With this change we use `file.asNio(eelProjectResidesOn)`.

`PackageCheckerBuildGradleKtsTemplateTest` and probably other tests fail on WSL and Docker (on Windows) due to this reason.

GitOrigin-RevId: 12dffacd43f444930d3eee5222ddd3f707daa852