[code-style] IJPL-213314 run formatting task monitoring coroutine explicitly on the default dispatcher
It is necessary so that the task coroutine can be properly monitored when single threaded dispatcher is used (i.e., under runBlocking in sync format).

In sync format, to modify the document, it is then necessary to switch back to the event loop of the runBlocking (if it is blocking the EDT, withContext(EDT) will deadlock).

GitOrigin-RevId: 6cab93724d1c3f88e32436186930a424efcc40e2