IJPL-190340 [eel, ijent] transition transferLocalContentToRemote to incremental transfer
For TransferTarget.Explicit, this change ensures directories are properly synced even when the target directory already exists. Previously, existing directories would be skipped. For TransferTarget.Temporary, incremental transfer eliminates the need for sourceHash, which relies on metadata. Instead, it traverses the directory and compares individual entries. Previously, the metadata-based hash would not reliably detect changes (for example, updating a file's contents is not reflected in the parent directory's metadata). Incremental transfer solves this issue, making the transfer accurate but slower.

GitOrigin-RevId: 33ef5ff658b6ca3e1f49eca7037fc1ae55c03382