[java-analysis] IDEA-373908 Freeze at com.intellij.codeInspection.dataFlow.CommonDataflow.getDataflowResult
Use ConcurrentFactoryMap instead of ConcurrentHashMap.computeIfAbsent
The ConcurrentHashMap.computeIfAbsent was used to prevent concurrent computations, as DFA can be computationally intensive and there are high chances of concurrent computations on the same method, so we didn't want to load all the cores with the same computation. However, blocking call cause deadlocks in rare cases in current threading model. Since CommonDataflow was introduced, we improved performance significantly and separated CFG creation into a separate cached action, so probably the CPU utilization is probably better now. In any case, it's better than a deadlock.

GitOrigin-RevId: 87951d83c9e51adddb34b49d0d8c75fa9201f79b