[RDCT] IJPL-212872 Fix data race condition in `TrafficLightRenderer#init`
In remote development without delays, sometimes `afterAdded` method is called for highlighters and then the same highlighters are processed in `for (rangeHighlighter in model.getAllHighlighters())` loop, leading to double number of errors/warnings in traffic light widget. This commit moves iteration and listener to the same thread, switching their order, to reduce possibility of such a race.

GitOrigin-RevId: d7ba8d848c799ed85fd48fd1fbf3c6e7ef95040b