[terminal] IJPL-211122 Add an ability to get shell startup options from TerminalView
Introduce the new `TerminalStartupOptions` interface for that. Decided to avoid using existing `ShellStartupOptions` because it is polluted with unrelated fields that are hard to remove due to external usages. So, let's use the new interface and consider further migration to it.

(cherry picked from commit 81e598bf5aec1a69c048e444e8eea5f425aef008)

IJ-CR-179105

GitOrigin-RevId: 534fd8daea6f1a3913e7328a78d397260e48d15f