PY-84617: Move WelcomeScreenLeftTabListener to the frontend
`WelcomeScreenFileType` is available only on the frontend module, so the listener must be executed on frontend as well.
Otherwise `event.newFile` is null and it is not possible to understand if the tab was changed to/from welcome screent tab.
Also, current implementation does not remove the `WelcomeScreenLeftPanel` but just change the pane.

GitOrigin-RevId: 124dfcd624dede4a4ee835461a03e3d7524e2ada