[terminal] IJPL-213971 Fix exception in inline completion TerminalOutputModelListener
The start offset of the block is not adjusted now when the content is trimmed. So, we need to always check offsets before accessing the model.
Fix it by using `getTypedCommandText` extension that perform all checks.
Also, this listener should be faster now, because we ignore changes that happen during command execution.

GitOrigin-RevId: 929782cd7191ea243c46b4997f9ad1f144d8048f