[UI] IJPL-189623 Fix icon positioning in SquareStripeButtonLook
The default implementation only takes the insets into account,
but not the icon padding. The insets are zero, so essentially
it just centers the icon. But we want the theme keys to affect
the icon position.

We could've fixed this by including the padding into the insets,
but it would be a risky fix, as there could be code that relies
on the insets being zero.

GitOrigin-RevId: 07a190d5236d630d5a5b5f8108dfa5081ef7335e