[UI, core] IJPL-213185 Simplify PsiDirectoryImpl.getBaseIcon
Because the base method always returns the placeholder
for directories (as getContainingFile is null for dirs),
we can safely just always return the base folder icon.

Fix the affected RubyMine test. It wasn't affected
by the previous fix because it uses a dummy icon manager
which creates a new icon instance every time,
so the "== NodePlaceholder" condition wasn't satisfied
and the placeholder was returned.

GitOrigin-RevId: e2c4d7cd20265450d3d64875d255c353dc87f768