[fleet] Ignore buildtool modules in JPS generation of test modules list
Why?
Aggregator tasks are a big machinery building a Fleet product to test against it.
Additionally, it requires the modules to be compliant with JPMS and their transitive dependencies too.

For example: transitive dependencies of the S3 AWS client are not compliant with JPMS.
It should not be a problem because it is never used in Fleet runtime.
But because it is used in test by that big machinery, it would be required to be compliant with JPMS, which is not good.

GitOrigin-RevId: 6e8ec2bffeba1078cbf36001b1daf0e7db8b1ee2