[terminal] IJPL-213148 Fix non-empty highlightings during update
Because the model is in an inconsistent state when
the document is being modified, we have a contract
that the highlighting snapshot will be empty at that moment.
This contract was violated during the latest refactoring.

Fix by extracting a function that wraps every modification.
Now it's guaranteed that the modification flag is set
and that the cursor position is corrected as needed.

Add a consistency test for this as well.

GitOrigin-RevId: 5ce7677c7e95a00c241fb96fd3c1a5dcafe120d9