[UI] IJPL-214014 Islands: update border colors
While we're at it, update Borders.ContrastBorderColor as well.
It doesn't seem to be used in the platform anymore,
but it has a lot of Android usages, and it seems a safe assumption
that they expect to have the same color.
Even its name suggests that it should be at least as contrast
as the regular border, so let's not make it worse at least.


(cherry picked from commit 7f23b431a84328ccdd72253b4e385dd2902892c6)

IJ-CR-179474

GitOrigin-RevId: 5afeae4a9dd0acdbd04f252414532938c565bf27