[UI] IJPL-214014 Islands: update border colors, 2nd attempt
This time do it with the alt key.

While we're at it, update Borders.ContrastBorderColor as well.
It doesn't seem to be used in the platform anymore,
but it has a lot of Android usages, and it seems a safe assumption
that they expect to have the same color.
Even its name suggests that it should be at least as contrast
as the regular border, so let's not make it worse at least.


(cherry picked from commit 2731acfabdd2429562afaf2e6f83ed57f322ace6)

IJ-CR-179474

GitOrigin-RevId: 373d47fb81cecd2f7fba6d46f710b8f770dff2d7