[terminal] IJPL-207841 Add text style prediction to type-ahead
Try to use the text style of the previous character for the next prediction.
It will improve the situation with blinking in PowerShell, but won't solve it fully. Because we still can't predict the style of the first character in the literal. So, very fast typed words will still appear with the default style, but then replaced with the actual one.

(cherry picked from commit 6e91cdea7d7236f592d5f3352adddf8e2e8f452f)

IJ-CR-179656

GitOrigin-RevId: e94cd2fad22bd583762486f10223b2d069b80070