[code-style] IJPL-213314 fix indicator may not be shown during async formatting
There is a common case where `FormattingTask.run` returns immediately after starting some `ProcessHandler`, but the progress was shown only for the duration of this call.

If `FormattingTask.isRunUnderProgress == true`, the progress indicator should be shown to the user until the `ProcessHandler` calls `AsyncFormattingRequest.onTextReady/onError`.

Which means that the coroutine that awaits `AsyncFormattingRequest.onTextReady/onError` must be wrapped in `withBackgroundProgress` and not just the `FormattingTask.run` itself.

GitOrigin-RevId: 7bd85c65044072116f1dba523fc22afb33e6e410