[multiverse] IJPL-201882 handle contexts on file move
When some file is moved to another directory, its contexts might change.
So, view providers associated with irrelevant contexts should be processed in some way.
The best way to deal with such view providers is just to invalidate them.
But PSI clients are not usually ready for that.
So let's try to preserve one of irrelevant view providers, clearing the context of this guy and setting its context to `any`.
It is possible to do that if there are no other relevant view providers for this file left, because it's not allowed to have `any` and definite context for a file simultaneously.

GitOrigin-RevId: 8abccccb947ebbabcab561001d05dda509a262d2