PY-84997: Do not display welcome project in the recent projects of WelcomeScreenLeftPanel
In PyCharm we are not hiding the welcome project since it might contain scripts created by a user.
So it should be possible to find the project, (see `PyCharmWelcomeScreenProjectProvider.doIsHiddenInRecentProjects`).
But at the same time, there is no reason to display welcome project in the resent projects inside the welcome project itself.

GitOrigin-RevId: e6aff912dadbf5a16e560999d147661447314dc4