[terminal] IJPL-214329 Fix the prompt separator position when command wraps to the next line
The problem is in the RangeHighlighter position.
Suppose it was created with bounds [0, 20] at the block creation. Then prompt text was added at offset 0 with length 10. And our highlighter is shifted to the range [10, 30] - so the separator starts painting in the wrong place.
Making it "greedy to left" fixes this problem and leaving the range highlighter at the start of the block. It should be safe, because any changes before the block start offset invalidate the block and create the new one.

But unfortunately, I failed to find why there was no such issue before.

(cherry picked from commit b5da26b2438104cf59fa1eb6461455dca6aa59e4)

IJ-CR-179741

GitOrigin-RevId: dc6cced7d7a4a644209f004beefb0b13131a5331