[rdct] IJPL-210428: Move frontend run configuration processing to background to avoid blocking call on EDT
Processing protocol events and registering configuration types on the EDT led to a blinking progress dialog during a blocking call. We decided to move configuration type processing to background. However, this broke the order of event processing on the frontend. We had to use a single-threaded dispatcher and move processing of other protocol events to background too to guarantee the order.

GitOrigin-RevId: cddf3998d476a90b723ff9450640b493a5e816c3