[rd debugger] IDEA-380440 Support performing `CreateRendererAction` in Split mode
The entire commit is about passing the correct presentation to `ActionLink`. This presentation contains user data (`BackendDelegatingAction.FRONTEND_UPDATE_RESULT`) that is absolutely necessary to decide where the action should be performed (backend/frontend). `AnActionLink` didn't allow passing a custom presentation so it created `AnActionEvent` with no user data, which forced `CreateRendererAction` from the popup to be performed on backend.

Alternative: allow `CreateRendererAction` to be performed on backend by making its `actionPerformed` dependent on IDE product mode. It'd mean supporting the action behavior in two modes, which we try to avoid. Plus it doesn't solve the problem with the other user data that could be as important in other scenarios.

GitOrigin-RevId: f5be119ca6698e4ceb735872b39bbf88f0808782