[run dashboard] IJPL-214553 Dirty hack with extension point to substitute frontend selection with backend run dashboard model entities
Since serialization lib is disallowed in api module, and half of dashboard code historically belongs there, we cannot simply access backend manager via the RunDashboardManagerProxy and resolve backend service by id - ID is serializable == inaccessible.

The current workaround for monolith is ugly and complicated, yet it solves the problem without reconfiguring the project or build tests. A proper implementation would be to get rid of SELECTED_ITEM data key usages completely across all service view-related actions. It requires a tremendous amount of work. We have to make sure that monolith is not broken, as well as split mode in new and fallback service view modes simultaneously because we abandoned the idea to split the run toolwindow.

GitOrigin-RevId: b78e29c9ac2b5776352230a038ab8c08dc550a11