[terminal] IJPL-214664 Fix cursor position blinking when printing right prompt
The issue is in the RangeHighlighter we install.
We install it after the cursor position change in the same EDT event. But then text may change in the next EDT event, move the highlighter and paint it.
Now we should update the highlighter after both text and cursor position change, so it should always paint in the right place.

GitOrigin-RevId: 32ea86dcde81233fc5e337133b8447ac2652e773