[terminal] IJPL-214664 Fix cursor position blinking when printing right prompt
The issue is in the RangeHighlighter we install.
We install it after the cursor position change in the same EDT event. But then text may change in the next EDT event, move the highlighter and paint it.
Now we should update the highlighter after both text and cursor position change, so it should always paint in the right place.


(cherry picked from commit 32ea86dcde81233fc5e337133b8447ac2652e773)

IJ-CR-179955

GitOrigin-RevId: c1229f9db3a841ca4f340811e8a3eee5f1f929bb