[terminal] IJPL-214664 Fix cursor blinking
Currently, the cursor highlighter is updated synchronously right after the text or cursor position change in the model. In the same EDT event.
But previous highlighter was removed asynchronously on painter job cancellation.
So, there was a situation when both new and old highlighter were painted on the screen at the same time. And it was looking like blinking.

The new approach is to remove the highlighter right before installing the new one - in the same EDT event with the model change.
It is kind of ugly because it is against of a coroutines idea, but I didn't find a better solution.

GitOrigin-RevId: 64826a1154b71acecfdb29ae8ef08869eb9b77c8