[platform UI] IJPL-189623 Fix the stripe button paddings
Turns out that it's needed to preserve the width of the hover area.
So instead of reducing the paddings, let's shift the whole thing
by adding some paddings to one side and removing the same
amount of padding from the other side.

The exact amounts were hand-picked by taking screenshots
and measuring the sizes. In some cases it's hard to center
exactly because there can be an odd number of pixels in total.
In such cases the solution that seems more centered was picked.


(cherry picked from commit ccc5cd71994020907e694c2e0dea41c54377b127)

IJ-CR-179392

GitOrigin-RevId: 58fc3199180501a645c306eda130ccfa142e8bd7