[UI] , Project View] IJPL-214537 Enable dir.expanded for remdev
It was initially added to measure only dir expansion,
to avoid polluting the statistics with non-dir nodes
that may expand faster / slower.

Let's replace that check with a marker interface
and enable it for all remdev nodes.
For remdev it isn't very important whether it's a dir or not,
because we're more interested in the effect of network latency.

(cherry picked from commit f39801aeae370a6ddcc94ea706d90778658eef00)

Reviewed and approved at IJ-CR-180090.

# Conflicts:
#	remote-dev/cwm-guest/src/com/jetbrains/thinclient/projectView/nodes/ThinClientProjectViewNode.kt

GitOrigin-RevId: 8d315d52fc8b4fe229e25109d290bc77f5348566