[UI] , Project View] IJPL-214537 Enable dir.expanded for remdev
It was initially added to measure only dir expansion,
to avoid polluting the statistics with non-dir nodes
that may expand faster / slower.

Let's replace that check with a marker interface
and enable it for all remdev nodes.
For remdev it isn't very important whether it's a dir or not,
because we're more interested in the effect of network latency.


(cherry picked from commit f39801aeae370a6ddcc94ea706d90778658eef00)

IJ-CR-180090

GitOrigin-RevId: d55e0cdf693fe079e503c4acbaeb75bc23265f5d