[kotlin] KT-80312: Preserve extrasBinaries when IDEA Serialization context is missing
Gradle Configuration Cache / Isolated Projects feature forces
Gradle Modules to be serialized & deserialized during configuration cache
serialization. To have consistent state for just stored cache (i.e. objects are live and hot)
and when it was loaded from store (i.e. objects were deserialized).
For example it will set all @Transient fields to null etc.

But this serialize & deserialize pipeline affects IdeaKotlinExtras because it depends on
`ideaKotlinSerializationContextOrNull` which is not available during model building.
Because it happens in Gradle process. So it gets serialized as Surrogate
with serialized entries in binary form (see usages of Companion.binaryExtras).
So at this moment data is preserved in a binary form.
And previous code was expecting that Surrogate will be deserialized only inside IDEA process
i.e. when models will be loading to be processed further in IDEA environment.

But with Configuration Cache / Project Isolation enabled Gradle will runs extra:
"serialize / deserialize" round.
And Surrogate's readResolve simply falls back to "emtpy" extras because
`ideaKotlinSerializationContextOrNull` is not available.

So the solution is to always preserve "binaryExtras" even when
ideaKotlinSerializationContextOrNull is not present.

And pipeline will look like this:

// Gradle Process, extras created by Gradle Plugins
myExtras = mutableExtras(myKey with myData)
// happens during model builder, when IDE invokes it
-> ideaExtras = "serialize myExtras and wrap it to IdeaExtras.from(serialized)""
// Now gradale does this "serialize/deserialize" round.
// And because ideaExtras implements java.io.Serializable
// configurationCacheSerializer will obey it
-> configurationCacheSerializer.serialize(ideaExtras) -> surrogateWithBinaryExtras
// and at this stage, previously it was returning empty "IdeaExtras", but now it preserves binaries
-> configurationCacheSerializer.deserialize(surrogateWithBinaryExtras)
---- Gradle <-> IDEA process crossing ---
-> serialize surrogateWithBinaryExtras + ideaKotlinSerializationContext
// get back initial state of myExtras with all custom keys and data
-> myExtras = mutableExtras(myKey with myData)

GitOrigin-RevId: 9d738585923df79b00f78257208a4542d38a267d