[collab/gh/gl] Disable navigation in per-commit reviews (IJPL-214779)
Navigation is simply not ready to work properly in these cases.
In most cases navigating from a focused thread to the next appears to work,
but navigating from an editor line to the next / previous comment doesn't work at all.

The reason is that editor lines are not mapped to the 'universal line' properly.

GitOrigin-RevId: b6508fc41b1d9599b4e08f8cc64a873af3bdef37