[find in files] IJPL-210278 Fix "Selected Files" and "Selected Directories" missing from scopes list
These two scopes depend on the data context and are shown only when multiple files or directories are selected.
Previously, all scopes were collected on the EDT right before the Find in Files dialog opened,
and the data context was correctly obtained in `CoroutineScopeModel.refreshScopes`.

After splitting FiF, scopes started being collected in the background, which delayed data context collection.
Because of this delay, the FiF dialog could open before the context was ready, and focus might shift,
causing the selected files or directories to be missing from the context. As a result,
the scopes list didn’t show “Selected Files” or “Selected Directories”.

The fix is to collect the data context as early as possible on the EDT and pass it to the async method.

GitOrigin-RevId: deb40d3e28078d366332b379a323edea36273106